requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai>=1.61.0
python-dotenv==1.0.1
pydantic>=2.9.0
python-multipart==0.0.9
pillow==10.2.0
